requirements.txt
torch==2.0.1
protobuf==3.20.2
transformers==4.18.0
jiwer
toml==0.10.2
librosa==0.9.1
soundfile == 0.10.2 
pandarallel==1.6.1
requests

python-Levenshtein==0.12.2
six==1.16.0
ctc-segmentation==1.7.1
inflect
loguru==0.7.0

onnx == 1.15.0 
onnxruntime-gpu == 1.15.1 
joblib == 1.1.0 
python-multipart


tritonclient[all]
fastapi
python-dotenv

uvicorn == 0.17.5
gunicorn == 20.1.0

pyctcdecode
https://github.com/kpu/kenlm/archive/master.zip
soundfile